스터디 신청 리스트 구현 완료 (서버 측 엔드포인트 재확인 필요)
