--- client/src/pages/MyStudyList.tsx
@@ -1,11 +1,0 @@
-
-
-const MyStudyList = () => {
-    return (
-        <div>
-            <h1>MyStudyList</h1>
-        </div>
-    );
-}
-
-export default MyStudyList;
