+++ client/src/pages/ProfileStudyList.tsx
@@ -1,0 +1,9 @@
+const ProfileStudyList = () => {
+  return (
+    <div>
+      <h1>MyStudyList</h1>
+    </div>
+  );
+};
+
+export default ProfileStudyList;
