+++ client/src/pages/StudyContent.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 // import { useState } from "react";
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 // import axios from "axios";
 import StudyComment from "../components/StudyComment";
